requirements.txt
leancloud~=2.9.12
fastapi~=0.89.1
pydantic~=1.10.4
python-dotenv~=0.21.1